backend/requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-storages==1.7.1
djangorestframework==3.12.4
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.7.2
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.6
